Move changing of email login data to auth resolver. Add rate limits
